+++ components/InfoIcon.svelte
@@ -1,1 +1,0 @@
-<i class="i-material-symbols-info-outline w-10"></i>