+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v2.2.0
+
+- Added domain allowlist/blocklist feature to control where the extension operates (thanks [@Aptcoder](https://github.com/Aptcoder)!)
+  - Users can specify domains where the extension should or should not run based on allow or block mode (default is allow)
+  - Supports wildcard domains (e.g. *.example.com)
+  - Supports specific overrides (e.g. if google.com is blocked, google.com/maps can be allowed)
+  - Adds color-coded status indicator showing whether name replacement is enabled or disabled on the current site
+- Changed extension icon to be more descriptive given the extension's state
+  - Has separate icons for the extension being enabled, disabled, and the current site being blocked
+- Moved the "Hide Debug Info" setting under an "Advanced Settings" dropdown section to keep settings more organized
+
 ## v2.1.0
 
 - Added keyboard shortcut functionality for quickly enabling/disabling the extension